lint

+++ ciftools/Binary/Encoding/Encoders/delta.py
@@ -2,7 +2,6 @@ import numpy as np
 from ciftools.binary.encoding.data_types import DataType, DataTypeEnum
 from ciftools.binary.encoding.encoders.base import CIFEncoderBase
 from ciftools.binary.encoding.encodings import DeltaEncoding, EncodingEnun
-
 from ciftools.binary.encoding.types import EncodedCIFData
 
 
@@ -11,7 +10,7 @@ class DeltaCIFEncoder(CIFEncoderBase):
         src_data_type: DataTypeEnum = DataType.from_dtype(data.dtype)
 
         if not src_data_type or src_data_type not in (DataTypeEnum.Int8, DataTypeEnum.Int16, DataTypeEnum.Int32):
-            data = data.astype(dtype='i4')
+            data = data.astype(dtype="i4")
             src_data_type = DataTypeEnum.Int32
 
         encoding: DeltaEncoding = {"kind": EncodingEnun.Delta, "srcType": src_data_type}
@@ -20,7 +19,7 @@ class DeltaCIFEncoder(CIFEncoderBase):
 
         if not data_length:
             encoding.origin = 0
-            return EncodedCIFData(data=np.empty(0, dtype='i4'), encoding=[encoding])
+            return EncodedCIFData(data=np.empty(0, dtype="i4"), encoding=[encoding])
 
         encoded_data = np.diff(data, prepend=data[0])
         encoding["origin"] = int(data[0])
